Adjusted logger interface and expose it.
