+++ src/platform/javascript/AlphaSynthWorkerSynthOutput.ts
@@ -1,7 +1,7 @@
 import { ISynthOutput } from '@src/synth/ISynthOutput';
 import { EventEmitter, IEventEmitter, IEventEmitterOfT, EventEmitterOfT } from '@src/EventEmitter';
 import { IWorkerScope } from '@src/platform/javascript/IWorkerScope';
-import { Logger } from '@src/util/Logger';
+import { Logger } from '@src/Logger';
 import { Environment } from '@src/Environment';
 
 /**
