+++ src/util/FontLoadingChecker.ts
@@ -1,5 +1,5 @@
 import { IEventEmitterOfT, EventEmitterOfT } from '@src/EventEmitter';
-import { Logger } from '@src/util/Logger';
+import { Logger } from '@src/Logger';
 import { Environment } from '@src/Environment';
 
 /**
