+++ test/audio/MidiPlaybackController.test.ts
@@ -2,7 +2,7 @@ import { MidiPlaybackController } from '@src/midi/MidiPlaybackController';
 import { AlphaTexImporter } from '@src/importer/AlphaTexImporter';
 import { Score } from '@src/model/Score';
 import { Settings } from '@src/Settings';
-import { Logger } from '@src/util/Logger';
+import { Logger } from '@src/Logger';
 import { GpImporterTestHelper } from '@test/importer/GpImporterTestHelper';
 import { TestPlatform } from '@test/TestPlatform';
